refactor: #43 monolith split — 3,143 → 784 lines across 17 focused files
Three-pass extraction of MobileRemoteShell monolith:
- Pass 1: Extract 10 component files + types + utils
- Pass 2: Extract effects, MediaLightbox, RuntimeBar, TopBar, SurfaceStatus
- Pass 3: Extract controller.ts (handlers + business logic)

Orchestrator now contains only state declarations and layout shell.
Clean typecheck, zero regressions, all functionality preserved.
